requirements.txt
google.genai 
flask
flask_jwt_extended
flask_sqlalchemy
python-dotenv
flask_cors
flask_bcrypt